Move CompatEnumNames to common so linter catches it
