+++ src/app/zap-templates/templates/app/cluster-enums.zapt
@@ -70,4 +70,4 @@ k{{asUpperCamelCase label}} = {{asHex mask}},
 } // namespace chip
 
 // Included at the end, so all our definitions above are available.
-#include <app/CompatEnumNames.h>
+#include <app/common/CompatEnumNames.h>
